+++ src/Ext/Techno/Body.cpp
@@ -598,6 +598,7 @@ void TechnoExt::ExtData::Serialize(T& Stm)
 		.Process(this->IsInTunnel)
 		.Process(this->DeployFireTimer)
 		.Process(this->ForceFullRearmDelay)
+		.Process(this->MobileRefineryTimer)
 		;
 }