+++ src/Ext/Techno/Hooks.cpp
@@ -29,6 +29,7 @@ DEFINE_HOOK(0x6F9E50, TechnoClass_AI, 0x5)
 	pExt->UpdateShield();
 	pExt->ApplySpawnLimitRange();
 	pExt->UpdateLaserTrails();
+	pExt->ApplyMobileRefinery();
 
 	TechnoExt::ApplyMindControlRangeLimit(pThis);