+++ src/Ext/TechnoType/Body.h
@@ -9,6 +9,7 @@
 #include <New/Type/LaserTrailTypeClass.h>
 #include <New/Type/Affiliated/InterceptorTypeClass.h>
 #include <New/Type/Affiliated/PassengerDeletionTypeClass.h>
+#include <New/Type/Affiliated/MobileRefineryTypeClass.h>
 #include <New/Type/DigitalDisplayTypeClass.h>
 
 class Matrix3D;
@@ -34,6 +35,8 @@ public:
 		Valueable<Leptons> MindControlRangeLimit;
 
 		std::unique_ptr<InterceptorTypeClass> InterceptorType;
+
+		std::unique_ptr<MobileRefineryTypeClass> MobileRefineryType;
 
 		Valueable<PartialVector3D<int>> TurretOffset;
 		Nullable<bool> TurretShadow;
@@ -198,6 +201,8 @@ public:
 			, MindControlRangeLimit {}
 
 			, InterceptorType { nullptr }
+
+			, MobileRefineryType { nullptr }
 
 			, TurretOffset { { 0, 0, 0 } }
 			, TurretShadow { }