+++ src/New/Type/Affiliated/MobileRefineryTypeClass.cpp
@@ -1,0 +1,62 @@
+#include "MobileRefineryTypeClass.h"
+
+#include <Utilities/SavegameDef.h>
+#include <Utilities/TemplateDef.h>
+
+MobileRefineryTypeClass::MobileRefineryTypeClass(TechnoTypeClass* OwnedBy) : OwnerType { OwnedBy }
+, TransDelay { 30 }
+, CashMultiplier { 1.0 }
+, AmountPerCell { 0 }
+, FrontOffset {}
+, LeftOffset {}
+, Display { true }
+, Display_Houses { AffectedHouse::All }
+, Anims {}
+, AnimMove { true }
+{ }
+
+void MobileRefineryTypeClass::LoadFromINI(CCINIClass* pINI, const char* pSection)
+{
+	INI_EX exINI(pINI);
+
+	this->TransDelay.Read(exINI, pSection, "MobileRefinery.TransDelay");
+	this->CashMultiplier.Read(exINI, pSection, "MobileRefinery.CashMultiplier");
+	this->AmountPerCell.Read(exINI, pSection, "MobileRefinery.AmountPerCell");
+	this->FrontOffset.Read(exINI, pSection, "MobileRefinery.FrontOffset");
+	this->LeftOffset.Read(exINI, pSection, "MobileRefinery.LeftOffset");
+	this->Display.Read(exINI, pSection, "MobileRefinery.Display");
+	this->Display_Houses.Read(exINI, pSection, "MobileRefinery.Display.Houses");
+	this->Anims.Read(exINI, pSection, "MobileRefinery.Anims");
+	this->AnimMove.Read(exINI, pSection, "MobileRefinery.AnimMove");
+}
+
+#pragma region(save/load)
+
+template <class T>
+bool MobileRefineryTypeClass::Serialize(T& stm)
+{
+	return stm
+		.Process(this->OwnerType)
+		.Process(this->TransDelay)
+		.Process(this->CashMultiplier)
+		.Process(this->AmountPerCell)
+		.Process(this->FrontOffset)
+		.Process(this->LeftOffset)
+		.Process(this->Display)
+		.Process(this->Display_Houses)
+		.Process(this->Anims)
+		.Process(this->AnimMove)
+		.Success();
+}
+
+bool MobileRefineryTypeClass::Load(PhobosStreamReader& stm, bool registerForChange)
+{
+	return this->Serialize(stm);
+}
+
+bool MobileRefineryTypeClass::Save(PhobosStreamWriter& stm) const
+{
+	return const_cast<MobileRefineryTypeClass*>(this)->Serialize(stm);
+}
+
+#pragma endregion(save/load)